Proper abstract job-store class, with file store implementation. Tests passing.
